0.1 Basic logic and style
Added:
Board to play(for now without X and O in canvas but in a script).

Logic for selecticting board fields, endgame, turns. Players cant select not empty field, and cant do anything when game is over.

Board is represented by 9 element array of new enum type(Empty, O, X).

Fader UI and board animation when game is turned on.